almost working

+++ Perceptron/src/Flower.java
@@ -1,23 +1,23 @@
 import java.util.Arrays;
 
 public class Flower implements Comparable<Flower>{
-    private double[] parameters;
+    private float[] parameters;
     private String flowerName;
 
-    public Flower(double[] parameters, String flowerName){
+    public Flower(float[] parameters, String flowerName){
         this.parameters = parameters;
         this.flowerName = flowerName;
     }
 
-    public Flower(double[] parameters) {
+    public Flower(float[] parameters) {
         this.parameters = parameters;
     }
 
-    public double[] getParameters() {
+    public float[] getParameters() {
         return parameters;
     }
 
-    public void setParameters(double[] parameters) {
+    public void setParameters(float[] parameters) {
         this.parameters = parameters;
     }
 
@@ -44,9 +44,9 @@ public class Flower implements Comparable<Flower>{
         } else return 0;
     }
 
-    public double sum(double[] params){
-        double sum = 0;
-        for(Double d : params){
+    public float sum(float[] params){
+        float sum = 0;
+        for(Float d : params){
             sum+=d;
         }
         return sum;
